* Cleaning and checking session while looking for a crashfix. * Make sure we have a valid map-session before pushing players to World. * Make sure both map- and world-session get deleted when socket gets disconnected.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@1059 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/MapMgr.cpp
@@ -250,62 +250,59 @@ void MapMgr::PushObject(ObjectPointer obj)
 		m_corpses.insert( TO_CORPSE(obj) );
 	}	
 	
-	obj->ClearInRangeSet();
-	ASSERT(obj->GetMapId() == _mapId);
-	if(!(obj->GetPositionX() < _maxX && obj->GetPositionX() > _minX) || 
-	   !(obj->GetPositionY() < _maxY && obj->GetPositionY() > _minY))
+	PlayerPointer plObj = NULLPLR;
+
+	if(obj->GetTypeId() == TYPEID_PLAYER)
 	{
-		if(obj->IsPlayer())
+		plObj = TO_PLAYER( obj );
+		if(plObj == NULL)
 		{
-			PlayerPointer plr = TO_PLAYER( obj );
-			if(plr->GetBindMapId() != GetMapId())
-			{
-				plr->SafeTeleport(plr->GetBindMapId(),0,plr->GetBindPositionX(),plr->GetBindPositionY(),plr->GetBindPositionZ(),0);
-				plr->GetSession()->SystemMessage("Teleported you to your hearthstone location as you were out of the map boundaries.");
-				return;
-			}
-			else
-			{
-				obj->GetPositionV()->ChangeCoords(plr->GetBindPositionX(),plr->GetBindPositionY(),plr->GetBindPositionZ(),0);
-				plr->GetSession()->SystemMessage("Teleported you to your hearthstone location as you were out of the map boundaries.");
-				WorldPacket * data = plr->BuildTeleportAckMsg(plr->GetPosition());
-				plr->GetSession()->SendPacket(data);
-				delete data;
-			}
+			Log.Error("MapMgr","Could not get a valid playerobject from object while trying to push to world");
+			return;
 		}
-		else
+		WorldSession * plSession = NULL;
+		plSession = plObj->GetSession();
+		if(plSession == NULL)
 		{
-			obj->GetPositionV()->ChangeCoords(0,0,0,0);
+			Log.Error("MapMgr","Could not get a valid session for player while trying to push to world");
+			return;
 		}
 	}
-
-	ASSERT(obj->GetPositionY() < _maxY && obj->GetPositionY() > _minY);
-	ASSERT(_cells);
+	
+	obj->ClearInRangeSet();
 
 	///////////////////////
 	// Get cell coordinates
 	///////////////////////
 
-	uint32 x = GetPosX(obj->GetPositionX());
-	uint32 y = GetPosY(obj->GetPositionY());
+	ASSERT(obj->GetMapId() == _mapId);
 
-	if(x >= _sizeX || y >= _sizeY)
+	ASSERT(obj->GetPositionY() < _maxY && obj->GetPositionY() > _minY);
+	ASSERT(_cells);
+
+	uint32 mx = obj->GetPositionX();
+	uint32 my = obj->GetPositionY();
+	uint32 cx = GetPosX(mx);
+	uint32 cy = GetPosY(my);
+
+	if(	mx > _maxX || my > _maxY ||
+		mx < _minX || my < _minY ||
+		cx >= _sizeX || cy >= _sizeY)
 	{
-		if(obj->IsPlayer())
+		if( plObj != NULL )
 		{
-			PlayerPointer plr = TO_PLAYER( obj );
-			if(plr->GetBindMapId() != GetMapId())
+			if(plObj->GetBindMapId() != GetMapId())
 			{
-				plr->SafeTeleport(plr->GetBindMapId(),0,plr->GetBindPositionX(),plr->GetBindPositionY(),plr->GetBindPositionZ(),0);
-				plr->GetSession()->SystemMessage("Teleported you to your hearthstone location as you were out of the map boundaries.");
+				plObj->SafeTeleport(plObj->GetBindMapId(),0,plObj->GetBindPositionX(),plObj->GetBindPositionY(),plObj->GetBindPositionZ(),0);
+				plObj->GetSession()->SystemMessage("Teleported you to your hearthstone location as you ended up on the wrong map.");
 				return;
 			}
 			else
 			{
-				obj->GetPositionV()->ChangeCoords(plr->GetBindPositionX(),plr->GetBindPositionY(),plr->GetBindPositionZ(),0);
-				plr->GetSession()->SystemMessage("Teleported you to your hearthstone location as you were out of the map boundaries.");
-				WorldPacket * data = plr->BuildTeleportAckMsg(plr->GetPosition());
-				plr->GetSession()->SendPacket(data);
+				obj->GetPositionV()->ChangeCoords(plObj->GetBindPositionX(),plObj->GetBindPositionY(),plObj->GetBindPositionZ(),0);
+				plObj->GetSession()->SystemMessage("Teleported you to your hearthstone location as you were out of the map boundaries.");
+				WorldPacket * data = plObj->BuildTeleportAckMsg(plObj->GetPosition());
+				plObj->GetSession()->SendPacket(data);
 				delete data;
 			}
 		}
@@ -314,32 +311,26 @@ void MapMgr::PushObject(ObjectPointer obj)
 			obj->GetPositionV()->ChangeCoords(0,0,0,0);
 		}
 
-		x = GetPosX(obj->GetPositionX());
-		y = GetPosY(obj->GetPositionY());
+		cx = GetPosX(obj->GetPositionX());
+		cy = GetPosY(obj->GetPositionY());
 	}
 
-	MapCell *objCell = GetCell(x,y);
+	MapCell *objCell = GetCell(cx,cy);
 	if (!objCell)
 	{
-		objCell = Create(x,y);
-		objCell->Init(x, y, _mapId, shared_from_this());
+		objCell = Create(cx,cy);
+		objCell->Init(cx, cy, _mapId, shared_from_this());
 	}
 
-	uint32 endX = (x <= _sizeX) ? x + 1 : (_sizeX-1);
-	uint32 endY = (y <= _sizeY) ? y + 1 : (_sizeY-1);
-	uint32 startX = x > 0 ? x - 1 : 0;
-	uint32 startY = y > 0 ? y - 1 : 0;
+	uint32 endX = (cx <= _sizeX) ? cx + 1 : (_sizeX-1);
+	uint32 endY = (cy <= _sizeY) ? cy + 1 : (_sizeY-1);
+	uint32 startX = cx > 0 ? cx - 1 : 0;
+	uint32 startY = cy > 0 ? cy - 1 : 0;
 	uint32 posX, posY;
 	MapCell *cell;
 	MapCell::ObjectSet::iterator iter;
 
 	uint32 count;
-	PlayerPointer plObj;
-
-	if(obj->GetTypeId() == TYPEID_PLAYER)
-		plObj = TO_PLAYER( obj );
-	else
-		plObj = NULLPLR;
 
 	if(plObj)
 	{
@@ -372,8 +363,8 @@ void MapMgr::PushObject(ObjectPointer obj)
 	 //Add to the mapmanager's object list
 	if(plObj)
 	{
-	   m_PlayerStorage[plObj->GetLowGUID()] = plObj;
-	   UpdateCellActivity(x, y, 2);
+		m_PlayerStorage[plObj->GetLowGUID()] = plObj;
+		UpdateCellActivity(cx, cy, 2);
 	}
 	else
 	{
@@ -1810,7 +1801,7 @@ void MapMgr::_PerformObjectDuties()
 
 	// Sessions are updated every loop.
 	{
-		int result;
+		int result = 0;
 		WorldSession * session;
 		SessionSet::iterator itr = Sessions.begin();
 		SessionSet::iterator it2;
@@ -1830,18 +1821,14 @@ void MapMgr::_PerformObjectDuties()
 			// Don't update players not on our map.
 			// If we abort in the handler, it means we will "lose" packets, or not process this.
 			// .. and that could be diasterous to our client :P
-			if(session->GetPlayer() && (!session->GetPlayer()->GetMapMgr() && session->GetPlayer()->GetMapMgr() != shared_from_this() ))
-			{
+			if( session->GetPlayer() != NULL && session->GetPlayer()->GetMapMgr()!= NULL && session->GetPlayer()->GetMapMgr() != shared_from_this())
 				continue;
-			}
 
-			if((result = session->Update(m_instanceID)))
+			result = session->Update(m_instanceID);
+			if(result)//session or socket deleted?
 			{
-				if(result == 1)
-				{
-					// complete deletion
+				if(result == 1)//socket don't exist anymore, delete from both world- and map-sessions.
 					sWorld.DeleteSession(session);
-				}
 				Sessions.erase(it2);
 			}
 		}
